:lock: V0.1 stiil have bug in GlobalBA()

+++ include/se2lam/MapPoint.h
@@ -35,7 +35,7 @@ public:
     void eraseObservation(const PtrKeyFrame& pKF);
     void eraseObservations(const std::vector<std::pair<PtrKeyFrame, size_t>>& obsCandidates);
     bool hasObservation(const PtrKeyFrame& pKF);
-    int countObservation();
+    size_t countObservation();
 
     int getOctave(const PtrKeyFrame pKF);
     size_t getFtrIdx(const PtrKeyFrame& pKF);
